requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
idna==3.1
jdcal==1.4.1
multidict==5.1.0
openpyxl==3.0.6
pydantic==1.7.3
python-dotenv==0.15.0
requests==2.25.1
tinvest==3.0.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
pytz~=2020.5
